Queue Jobs moved into different namespace
